Improve program list translations (#554)
